Add main.py with Book class and main function
I added a new file called main.py to the repo. In this file, I created
a class called book which encapsulates the functionality related to
processing books.

Within the Book class, I have created an initializer method that takes a
file path as a parameter. This method initializes several properties to
store information about the book, including:

    - path: the file path of the book
    - text_content: the content of the book obtained by reading the
      specified file
    - word_count: the number of words in the book's text content
    - char_counts: a dictionary storing the frequency of each character
      in the book's text content

To process the book data and populate these properties, I have
implemented two private methods:

    - __get_text_content(): This method reads and retrieves the text
      content from the specified file.
    from the specified file.
    - __get_char_count(): This method calculates the frequency of each
      character in the text content and stores it in the char_counts
      property.

I chose to make these methods private because they are intended for
internal use within the Book class and are not meant to be accessed
externally.

To demonstrate the usage of the Book class, I defined a main function
that creates an instance of the Book class. In this case, I instantiated
the Book class with the file path "books/frankenstein.txt". This allows
us to analyze the "Frankenstein" book specifically.